update vendorName to vendorId, split users page into separate components
